[#2103] Refactor `parser` package for greater organisation of classes (#2104)
Classes in the `parser` package as of now are not properly organised,
with exceptions, types and parser classes being found in the same
directory level.

This does not follow the convention established by other packages,
whereby logically similar classes are further grouped into
sub-packages.

Let's move to refactor the `parser` package to ensure that it follows
the overall structure of the other packages and to increase the level
of organisation within the `parser` package.
