+++ src/main/java/reposense/RepoSense.java
@@ -15,9 +15,9 @@ import reposense.model.RepoConfiguration;
 import reposense.model.ReportConfiguration;
 import reposense.model.RunConfigurationDecider;
 import reposense.parser.ArgsParser;
-import reposense.parser.InvalidCsvException;
-import reposense.parser.InvalidHeaderException;
-import reposense.parser.ParseException;
+import reposense.parser.exceptions.InvalidCsvException;
+import reposense.parser.exceptions.InvalidHeaderException;
+import reposense.parser.exceptions.ParseException;
 import reposense.report.ReportGenerator;
 import reposense.system.LogsManager;
 import reposense.system.ReportServer;
