+++ src/main/java/reposense/model/CliRunConfiguration.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import reposense.parser.InvalidLocationException;
-import reposense.parser.ParseException;
+import reposense.parser.exceptions.InvalidLocationException;
+import reposense.parser.exceptions.ParseException;
 import reposense.system.LogsManager;
 
 /**
