+++ src/main/java/reposense/model/ConfigRunConfiguration.java
@@ -9,9 +9,9 @@ import java.util.logging.Logger;
 
 import reposense.parser.AuthorConfigCsvParser;
 import reposense.parser.GroupConfigCsvParser;
-import reposense.parser.InvalidCsvException;
-import reposense.parser.InvalidHeaderException;
 import reposense.parser.RepoConfigCsvParser;
+import reposense.parser.exceptions.InvalidCsvException;
+import reposense.parser.exceptions.InvalidHeaderException;
 import reposense.system.LogsManager;
 
 /**
