+++ src/main/java/reposense/model/RunConfiguration.java
@@ -3,9 +3,9 @@ package reposense.model;
 import java.io.IOException;
 import java.util.List;
 
-import reposense.parser.InvalidCsvException;
-import reposense.parser.InvalidHeaderException;
-import reposense.parser.ParseException;
+import reposense.parser.exceptions.InvalidCsvException;
+import reposense.parser.exceptions.InvalidHeaderException;
+import reposense.parser.exceptions.ParseException;
 
 /**
  * Interface to get configurations for current RepoSense run.
