+++ src/main/java/reposense/parser/ArgsParser.java
@@ -28,6 +28,18 @@ import reposense.RepoSense;
 import reposense.model.CliArguments;
 import reposense.model.FileType;
 import reposense.model.ReportConfiguration;
+import reposense.parser.exceptions.ParseException;
+import reposense.parser.types.AlphanumericArgumentType;
+import reposense.parser.types.AnalysisThreadsArgumentType;
+import reposense.parser.types.AssetsFolderArgumentType;
+import reposense.parser.types.CloningThreadsArgumentType;
+import reposense.parser.types.ConfigFolderArgumentType;
+import reposense.parser.types.OutputFolderArgumentType;
+import reposense.parser.types.PeriodArgumentType;
+import reposense.parser.types.ReportFolderArgumentType;
+import reposense.parser.types.SinceDateArgumentType;
+import reposense.parser.types.UntilDateArgumentType;
+import reposense.parser.types.ZoneIdArgumentType;
 import reposense.system.LogsManager;
 import reposense.util.TimeUtil;
