+++ src/main/java/reposense/parser/types/AlphanumericArgumentType.java
@@ -1,4 +1,4 @@
-package reposense.parser;
+package reposense.parser.types;
 
 import java.util.regex.Pattern;
