+++ src/main/java/reposense/parser/types/OutputFolderArgumentType.java
@@ -1,4 +1,4 @@
-package reposense.parser;
+package reposense.parser.types;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -8,15 +8,18 @@ import net.sourceforge.argparse4j.inf.Argument;
 import net.sourceforge.argparse4j.inf.ArgumentParser;
 import net.sourceforge.argparse4j.inf.ArgumentParserException;
 import net.sourceforge.argparse4j.inf.ArgumentType;
+import reposense.parser.ArgsParser;
 
 /**
- * Checks the argument of {@code --assets} flag.
+ * Checks the argument of {@code --output} flag.
  */
-public class AssetsFolderArgumentType implements ArgumentType<Path> {
+public class OutputFolderArgumentType implements ArgumentType<Path> {
     @Override
     public Path convert(ArgumentParser parser, Argument arg, String value) throws ArgumentParserException {
         // Piggyback on library methods to do file existence checks
-        Arguments.fileType().verifyExists().verifyIsDirectory().verifyCanRead().convert(parser, arg, value);
-        return Paths.get(value);
+        Arguments.fileType().verifyExists().verifyIsDirectory().verifyCanWrite()
+                .or()
+                .verifyNotExists().convert(parser, arg, value);
+        return Paths.get(value).resolve(ArgsParser.DEFAULT_REPORT_NAME);
     }
 }
