+++ src/test/java/reposense/parser/ArgsParserTest.java
@@ -26,6 +26,8 @@ import reposense.model.FileType;
 import reposense.model.FileTypeTest;
 import reposense.model.RepoConfiguration;
 import reposense.model.RunConfigurationDecider;
+import reposense.parser.exceptions.ParseException;
+import reposense.parser.types.SinceDateArgumentType;
 import reposense.util.FileUtil;
 import reposense.util.InputBuilder;
 import reposense.util.TestUtil;
