+++ src/test/java/reposense/parser/RepoConfigParserTest.java
@@ -19,6 +19,8 @@ import reposense.model.CommitHash;
 import reposense.model.FileType;
 import reposense.model.RepoConfiguration;
 import reposense.model.RepoLocation;
+import reposense.parser.exceptions.InvalidCsvException;
+import reposense.parser.exceptions.InvalidHeaderException;
 import reposense.util.InputBuilder;
 import reposense.util.TestUtil;
